Dialog

+++ src/app/favourite/favorites.component.ts
@@ -7,6 +7,7 @@ import { Toasty, } from 'nativescript-toasty';
 import { View } from 'tns-core-modules/ui/page';
 import { RadListViewComponent } from 'nativescript-ui-listview/angular';
 import { ListViewEventData } from 'nativescript-ui-listview';
+import { confirm } from "tns-core-modules/ui/dialogs";
 // import { View } from 'ui/core/view';
 @Component({
     selector: 'app-favorites',
@@ -36,10 +37,35 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
     }
 
     deleteFavorite(id: number) {
-        this.favoriteservice.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
-                errmess => this.errMess = errmess);
-    }
+        console.log('delete', id);
+    
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish '+ id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        };
+    
+        confirm(options).then((result: boolean) => {
+            if(result) {
+    
+              this.favorites = null;
+    
+              this.favoriteservice.deleteFavorite(id)
+                  .subscribe(favorites => { 
+                    const toast = new Toasty("Deleted Dish "+ id, "short", "bottom");
+                    toast.show();
+                    this.favorites = new ObservableArray(favorites);
+                  },
+                  errmess => this.errMess = errmess);
+            }
+            else {
+              console.log('Delete cancelled');
+            }
+        });
+    
+      }
 
     public onCellSwiping(args: ListViewEventData) {
         var swipeLimits = args.data.swipeLimits;
